kernel-ffi: split Qwen3.6 MoE host wrappers
