+++ crates/kernel-ffi/src/qwen36_moe/prefill.rs
@@ -1,2311 +1,3 @@
-//! FFI bridge for the Qwen3.6-MoE persistent decode megakernel.
-//!
-//! Status: **infrastructure only.** PR 4 (this file) lands the descriptor
-//! layout + a stub kernel that validates the launch path and descriptor
-//! read-out. Actual transformer compute (attention with `attn_output_gate`,
-//! MoE routing + work-stealing dispatch, fused expert mat-vec, shared
-//! expert, lm_head) lands in follow-up PRs.
-//!
-//! Why a stub first: the megakernel will be ~6500 LoC of HIP. Wiring the
-//! FFI/bridge/build path before any compute math means a future kernel
-//! commit can land focused-on-math code with no orchestration noise. The
-//! stub also serves as a smoke test for descriptor field layout — a
-//! Rust↔C++ struct-mismatch bug found here saves hours later.
-
-#![cfg_attr(
-    not(supersonic_backend_hip),
-    allow(unused_variables, unused_mut, unreachable_code)
-)]
-
-use std::collections::{HashMap, VecDeque};
-use std::ffi::{c_int, c_void};
-use std::os::raw::c_uint;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::{Arc, Condvar, Mutex, OnceLock};
-
-use gpu_hal::{Backend, BufferKind, GpuBuffer, GpuError, ScalarType};
-use half::f16;
-
-use crate::layer_desc::MAX_BATCH_SIZE;
-
-const QWEN36_ROUTE_PROFILE_DEFAULT_LAYERS: usize = 40;
-const QWEN36_ROUTE_PROFILE_DEFAULT_MAX_CALLS: usize = 16_384;
-const QWEN36_ROUTE_PROFILE_DEFAULT_CAPS: [usize; 6] = [2, 4, 8, 16, 32, 64];
-const QWEN36_BATCHED_PREFILL_PLAN_CHUNKS: [usize; 5] = [64, 128, 256, 512, 1024];
-
-static QWEN36_ROUTE_PROFILE: OnceLock<Mutex<Qwen36RouteProfileAccumulator>> = OnceLock::new();
-static QWEN36_EXPERT_RESIDENCY_PROFILE: OnceLock<Mutex<Qwen36ExpertResidencyProfileAccumulator>> =
-    OnceLock::new();
-static QWEN36_BATCHED_PREFILL_FEASIBILITY_PROFILE: OnceLock<
-    Mutex<Qwen36BatchedPrefillFeasibilityConfig>,
-> = OnceLock::new();
-static QWEN36_FFN_ROUTER_PARITY_TAP_CALLS: AtomicUsize = AtomicUsize::new(0);
-static QWEN36_FFN_SHARED_PARITY_TAP_CALLS: AtomicUsize = AtomicUsize::new(0);
-
-#[derive(Debug, Clone, Default)]
-struct Qwen36RouteProfileAccumulator {
-    records: Vec<Vec<u16>>,
-    dropped_calls: u64,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36RouteProfileSnapshot {
-    pub calls: u64,
-    pub dropped_calls: u64,
-    pub layers: usize,
-    pub assignments: u64,
-    pub unique_layer_experts: usize,
-    pub adjacent_hits: u64,
-    pub adjacent_total: u64,
-    pub cache_sims: Vec<Qwen36RouteCacheSim>,
-    pub topn_sims: Vec<Qwen36RouteTopNSim>,
-    pub topn_layers: Vec<Qwen36RouteTopNLayer>,
-    pub route_calls: Vec<Qwen36RouteCall>,
-}
-
-impl Qwen36RouteProfileSnapshot {
-    pub fn adjacent_hit_rate(&self) -> f64 {
-        if self.adjacent_total == 0 {
-            0.0
-        } else {
-            self.adjacent_hits as f64 / self.adjacent_total as f64
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36RouteCacheSim {
-    pub capacity: usize,
-    pub hits: u64,
-    pub misses: u64,
-}
-
-impl Qwen36RouteCacheSim {
-    pub fn hit_rate(&self) -> f64 {
-        let total = self.hits + self.misses;
-        if total == 0 {
-            0.0
-        } else {
-            self.hits as f64 / total as f64
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36RouteTopNSim {
-    pub capacity: usize,
-    pub covered: u64,
-    pub total: u64,
-}
-
-impl Qwen36RouteTopNSim {
-    pub fn coverage(&self) -> f64 {
-        if self.total == 0 {
-            0.0
-        } else {
-            self.covered as f64 / self.total as f64
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36RouteTopNLayer {
-    pub capacity: usize,
-    pub layer: usize,
-    pub experts: Vec<u16>,
-    pub counts: Vec<u64>,
-    pub covered: u64,
-    pub total: u64,
-}
-
-impl Qwen36RouteTopNLayer {
-    pub fn coverage(&self) -> f64 {
-        if self.total == 0 {
-            0.0
-        } else {
-            self.covered as f64 / self.total as f64
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36RouteCall {
-    pub call_idx: usize,
-    pub layer: usize,
-    pub experts: Vec<u16>,
-}
-
-#[derive(Debug, Clone)]
-struct Qwen36BatchedPrefillFeasibilityConfig {
-    layers: usize,
-    top_k: usize,
-    num_experts: usize,
-    chunk_size: usize,
-    prefill_tokens: usize,
-}
-
-impl Default for Qwen36BatchedPrefillFeasibilityConfig {
-    fn default() -> Self {
-        Self {
-            layers: QWEN36_ROUTE_PROFILE_DEFAULT_LAYERS,
-            top_k: 8,
-            num_experts: 256,
-            chunk_size: 512,
-            prefill_tokens: 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36BatchedPrefillFeasibilityProfileSnapshot {
-    pub calls: u64,
-    pub dropped_calls: u64,
-    pub layers: usize,
-    pub top_k: usize,
-    pub num_experts: usize,
-    pub chunk_size: usize,
-    pub prefill_tokens: usize,
-    pub profiled_tokens: usize,
-    pub chunks: usize,
-    pub assignments: u64,
-    pub permutation_entries: u64,
-    pub expert_segments: u64,
-    pub wmma16_segments: u64,
-    pub wmma16_covered_assignments: u64,
-    pub wmma16_padded_assignments: u64,
-    pub scalar_tail_segments: u64,
-    pub scalar_tail_assignments: u64,
-    pub max_rows_per_segment: u64,
-    pub max_unique_experts_per_layer_chunk: usize,
-}
-
-impl Qwen36BatchedPrefillFeasibilityProfileSnapshot {
-    pub fn avg_unique_experts_per_layer_chunk(&self) -> f64 {
-        let denom = self.chunks.saturating_mul(self.layers);
-        if denom == 0 {
-            0.0
-        } else {
-            self.expert_segments as f64 / denom as f64
-        }
-    }
-
-    pub fn avg_rows_per_segment(&self) -> f64 {
-        if self.expert_segments == 0 {
-            0.0
-        } else {
-            self.assignments as f64 / self.expert_segments as f64
-        }
-    }
-
-    pub fn wmma16_assignment_coverage(&self) -> f64 {
-        if self.assignments == 0 {
-            0.0
-        } else {
-            self.wmma16_covered_assignments as f64 / self.assignments as f64
-        }
-    }
-
-    pub fn wmma16_padding_overhead(&self) -> f64 {
-        if self.assignments == 0 {
-            0.0
-        } else {
-            self.wmma16_padded_assignments
-                .saturating_sub(self.assignments) as f64
-                / self.assignments as f64
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Qwen36ExpertResidentFormat {
-    NativeInt4,
-    Fp16Mps,
-}
-
-impl Qwen36ExpertResidentFormat {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::NativeInt4 => "native_int4",
-            Self::Fp16Mps => "fp16_mps",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Qwen36ExpertResidencyScope {
-    PerLayer,
-}
-
-impl Qwen36ExpertResidencyScope {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::PerLayer => "per_layer",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Qwen36ExpertResidencyMissPolicy {
-    ExactRoute,
-    LruHotset,
-    GpuPack,
-    StaticTopN,
-}
-
-impl Qwen36ExpertResidencyMissPolicy {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::ExactRoute => "exact_route",
-            Self::LruHotset => "lru_hotset",
-            Self::GpuPack => "gpu_pack",
-            Self::StaticTopN => "static_topn",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct Qwen36ExpertResidencyProfileKey {
-    pub resident_format: Qwen36ExpertResidentFormat,
-    pub scope: Qwen36ExpertResidencyScope,
-    pub miss_policy: Qwen36ExpertResidencyMissPolicy,
-    pub capacity: usize,
-}
-
-impl Qwen36ExpertResidencyProfileKey {
-    fn new(miss_policy: Qwen36ExpertResidencyMissPolicy, capacity: usize) -> Self {
-        Self::with_format(
-            Qwen36ExpertResidentFormat::NativeInt4,
-            miss_policy,
-            capacity,
-        )
-    }
-
-    fn with_format(
-        resident_format: Qwen36ExpertResidentFormat,
-        miss_policy: Qwen36ExpertResidencyMissPolicy,
-        capacity: usize,
-    ) -> Self {
-        Self {
-            resident_format,
-            scope: Qwen36ExpertResidencyScope::PerLayer,
-            miss_policy,
-            capacity,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-struct Qwen36ExpertResidencyCounters {
-    calls: u64,
-    exact_hits: u64,
-    route_refills: u64,
-    allocations: u64,
-    copied_bytes: u64,
-    active_groups_total: u64,
-    max_active_groups: usize,
-    slot_hits: u64,
-    slot_misses: u64,
-    evictions: u64,
-}
-
-#[derive(Debug, Clone, Default)]
-struct Qwen36ExpertResidencyProfileAccumulator {
-    totals: Qwen36ExpertResidencyCounters,
-    policies: HashMap<Qwen36ExpertResidencyProfileKey, Qwen36ExpertResidencyCounters>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36ExpertResidencyPolicySnapshot {
-    pub resident_format: &'static str,
-    pub scope: &'static str,
-    pub miss_policy: &'static str,
-    pub capacity: usize,
-    pub calls: u64,
-    pub exact_hits: u64,
-    pub route_refills: u64,
-    pub allocations: u64,
-    pub copied_bytes: u64,
-    pub active_groups_total: u64,
-    pub max_active_groups: usize,
-    pub slot_hits: u64,
-    pub slot_misses: u64,
-    pub evictions: u64,
-}
-
-impl Qwen36ExpertResidencyPolicySnapshot {
-    pub fn exact_hit_rate(&self) -> f64 {
-        if self.calls == 0 {
-            0.0
-        } else {
-            self.exact_hits as f64 / self.calls as f64
-        }
-    }
-
-    pub fn avg_active_groups(&self) -> f64 {
-        if self.calls == 0 {
-            0.0
-        } else {
-            self.active_groups_total as f64 / self.calls as f64
-        }
-    }
-
-    pub fn avg_copied_bytes(&self) -> f64 {
-        let copy_ops = self.route_refills + self.allocations;
-        if copy_ops == 0 {
-            0.0
-        } else {
-            self.copied_bytes as f64 / copy_ops as f64
-        }
-    }
-
-    pub fn slot_hit_rate(&self) -> f64 {
-        let total = self.slot_hits + self.slot_misses;
-        if total == 0 {
-            0.0
-        } else {
-            self.slot_hits as f64 / total as f64
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Qwen36ExpertResidencyProfileSnapshot {
-    pub calls: u64,
-    pub exact_hits: u64,
-    pub route_refills: u64,
-    pub allocations: u64,
-    pub copied_bytes: u64,
-    pub active_groups_total: u64,
-    pub max_active_groups: usize,
-    pub entries: usize,
-    pub slot_hits: u64,
-    pub slot_misses: u64,
-    pub evictions: u64,
-    pub policies: Vec<Qwen36ExpertResidencyPolicySnapshot>,
-}
-
-impl Qwen36ExpertResidencyProfileSnapshot {
-    pub fn exact_hit_rate(&self) -> f64 {
-        if self.calls == 0 {
-            0.0
-        } else {
-            self.exact_hits as f64 / self.calls as f64
-        }
-    }
-
-    pub fn avg_active_groups(&self) -> f64 {
-        if self.calls == 0 {
-            0.0
-        } else {
-            self.active_groups_total as f64 / self.calls as f64
-        }
-    }
-
-    pub fn avg_copied_bytes(&self) -> f64 {
-        let copy_ops = self.route_refills + self.allocations;
-        if copy_ops == 0 {
-            0.0
-        } else {
-            self.copied_bytes as f64 / copy_ops as f64
-        }
-    }
-
-    pub fn slot_hit_rate(&self) -> f64 {
-        let total = self.slot_hits + self.slot_misses;
-        if total == 0 {
-            0.0
-        } else {
-            self.slot_hits as f64 / total as f64
-        }
-    }
-}
-
-pub type Qwen36PackedExpertCacheProfileSnapshot = Qwen36ExpertResidencyProfileSnapshot;
-
-#[allow(clippy::too_many_arguments)]
-fn qwen36_expert_residency_counters_record(
-    counters: &mut Qwen36ExpertResidencyCounters,
-    exact_hit: bool,
-    route_refill: bool,
-    allocation: bool,
-    active_groups: usize,
-    copied_bytes: usize,
-    slot_hits: usize,
-    slot_misses: usize,
-    evictions: usize,
-) {
-    counters.calls += 1;
-    counters.active_groups_total += active_groups as u64;
-    counters.max_active_groups = counters.max_active_groups.max(active_groups);
-    if exact_hit {
-        counters.exact_hits += 1;
-    }
-    if route_refill {
-        counters.route_refills += 1;
-    }
-    if allocation {
-        counters.allocations += 1;
-    }
-    counters.copied_bytes += copied_bytes as u64;
-    counters.slot_hits += slot_hits as u64;
-    counters.slot_misses += slot_misses as u64;
-    counters.evictions += evictions as u64;
-}
-
-pub fn qwen36_route_profile_reset() {
-    if let Some(profile) = QWEN36_ROUTE_PROFILE.get() {
-        let mut profile = profile.lock().expect("qwen36 route profile mutex poisoned");
-        profile.records.clear();
-        profile.dropped_calls = 0;
-    }
-}
-
-pub fn qwen36_batched_prefill_feasibility_profile_reset() {
-    if let Some(profile) = QWEN36_BATCHED_PREFILL_FEASIBILITY_PROFILE.get() {
-        let mut profile = profile
-            .lock()
-            .expect("qwen36 batched prefill feasibility profile mutex poisoned");
-        *profile = Qwen36BatchedPrefillFeasibilityConfig::default();
-    }
-}
-
-pub fn qwen36_route_profile_snapshot() -> Qwen36RouteProfileSnapshot {
-    let Some(profile) = QWEN36_ROUTE_PROFILE.get() else {
-        return Qwen36RouteProfileSnapshot::default();
-    };
-    let profile = profile.lock().expect("qwen36 route profile mutex poisoned");
-    qwen36_route_profile_simulate(
-        &profile.records,
-        profile.dropped_calls,
-        qwen36_route_profile_layers(),
-    )
-}
-
-pub fn qwen36_batched_prefill_feasibility_profile_configure(
-    layers: usize,
-    top_k: usize,
-    num_experts: usize,
-    chunk_size: usize,
-    prefill_tokens: usize,
-) {
-    let profile = QWEN36_BATCHED_PREFILL_FEASIBILITY_PROFILE
-        .get_or_init(|| Mutex::new(Qwen36BatchedPrefillFeasibilityConfig::default()));
-    let mut profile = profile
-        .lock()
-        .expect("qwen36 batched prefill feasibility profile mutex poisoned");
-    *profile = Qwen36BatchedPrefillFeasibilityConfig {
-        layers: layers.max(1),
-        top_k: top_k.max(1),
-        num_experts: num_experts.max(1),
-        chunk_size: chunk_size.max(1),
-        prefill_tokens,
-    };
-}
-
-pub fn qwen36_batched_prefill_feasibility_profile_snapshot(
-) -> Qwen36BatchedPrefillFeasibilityProfileSnapshot {
-    let Some(route_profile) = QWEN36_ROUTE_PROFILE.get() else {
-        return Qwen36BatchedPrefillFeasibilityProfileSnapshot::default();
-    };
-    let route_profile = route_profile
-        .lock()
-        .expect("qwen36 route profile mutex poisoned");
-    let config = QWEN36_BATCHED_PREFILL_FEASIBILITY_PROFILE
-        .get()
-        .map(|profile| {
-            profile
-                .lock()
-                .expect("qwen36 batched prefill feasibility profile mutex poisoned")
-                .clone()
-        })
-        .unwrap_or_default();
-    qwen36_batched_prefill_feasibility_profile_simulate(
-        &route_profile.records,
-        route_profile.dropped_calls,
-        &config,
-    )
-}
-
-pub fn qwen36_batched_prefill_feasibility_plan_snapshots(
-) -> Vec<Qwen36BatchedPrefillFeasibilityProfileSnapshot> {
-    let Some(route_profile) = QWEN36_ROUTE_PROFILE.get() else {
-        return Vec::new();
-    };
-    let route_profile = route_profile
-        .lock()
-        .expect("qwen36 route profile mutex poisoned");
-    let config = QWEN36_BATCHED_PREFILL_FEASIBILITY_PROFILE
-        .get()
-        .map(|profile| {
-            profile
-                .lock()
-                .expect("qwen36 batched prefill feasibility profile mutex poisoned")
-                .clone()
-        })
-        .unwrap_or_default();
-    let mut chunk_sizes = qwen36_batched_prefill_plan_chunk_sizes(config.chunk_size);
-    chunk_sizes.sort_unstable();
-    chunk_sizes.dedup();
-    chunk_sizes
-        .into_iter()
-        .map(|chunk_size| {
-            let mut plan_config = config.clone();
-            plan_config.chunk_size = chunk_size.max(1);
-            qwen36_batched_prefill_feasibility_profile_simulate(
-                &route_profile.records,
-                route_profile.dropped_calls,
-                &plan_config,
-            )
-        })
-        .collect()
-}
-
-pub fn qwen36_packed_expert_cache_profile_reset() {
-    qwen36_expert_residency_profile_reset();
-}
-
-pub fn qwen36_expert_residency_profile_reset() {
-    if let Some(profile) = QWEN36_EXPERT_RESIDENCY_PROFILE.get() {
-        *profile
-            .lock()
-            .expect("qwen36 expert residency profile mutex poisoned") =
-            Qwen36ExpertResidencyProfileAccumulator::default();
-    }
-}
-
-pub fn qwen36_packed_expert_cache_profile_snapshot() -> Qwen36PackedExpertCacheProfileSnapshot {
-    qwen36_expert_residency_profile_snapshot()
-}
-
-pub fn qwen36_expert_residency_profile_snapshot() -> Qwen36ExpertResidencyProfileSnapshot {
-    let entries = QWEN36_PACKED_EXPERT_CACHE
-        .get()
-        .and_then(|cache| cache.lock().ok().map(|cache| cache.len()))
-        .unwrap_or(0)
-        + QWEN36_PACKED_EXPERT_HOTSET_CACHE
-            .get()
-            .and_then(|cache| cache.lock().ok().map(|cache| cache.len()))
-            .unwrap_or(0)
-        + QWEN36_PACKED_EXPERT_STATIC_TOPN_CACHE
-            .get()
-            .and_then(|cache| cache.lock().ok().map(|cache| cache.len()))
-            .unwrap_or(0)
-        + QWEN36_MPS_EXPERT_STATIC_TOPN_CACHE
-            .get()
-            .and_then(|cache| cache.lock().ok().map(|cache| cache.len()))
-            .unwrap_or(0);
-    let Some(profile) = QWEN36_EXPERT_RESIDENCY_PROFILE.get() else {
-        return Qwen36ExpertResidencyProfileSnapshot {
-            entries,
-            ..Qwen36ExpertResidencyProfileSnapshot::default()
-        };
-    };
-    let profile = profile
-        .lock()
-        .expect("qwen36 expert residency profile mutex poisoned");
-    let mut policies: Vec<_> = profile
-        .policies
-        .iter()
-        .map(|(key, counters)| Qwen36ExpertResidencyPolicySnapshot {
-            resident_format: key.resident_format.as_str(),
-            scope: key.scope.as_str(),
-            miss_policy: key.miss_policy.as_str(),
-            capacity: key.capacity,
-            calls: counters.calls,
-            exact_hits: counters.exact_hits,
-            route_refills: counters.route_refills,
-            allocations: counters.allocations,
-            copied_bytes: counters.copied_bytes,
-            active_groups_total: counters.active_groups_total,
-            max_active_groups: counters.max_active_groups,
-            slot_hits: counters.slot_hits,
-            slot_misses: counters.slot_misses,
-            evictions: counters.evictions,
-        })
-        .collect();
-    policies.sort_by_key(|policy| {
-        (
-            policy.resident_format,
-            policy.scope,
-            policy.miss_policy,
-            policy.capacity,
-        )
-    });
-    Qwen36ExpertResidencyProfileSnapshot {
-        calls: profile.totals.calls,
-        exact_hits: profile.totals.exact_hits,
-        route_refills: profile.totals.route_refills,
-        allocations: profile.totals.allocations,
-        copied_bytes: profile.totals.copied_bytes,
-        active_groups_total: profile.totals.active_groups_total,
-        max_active_groups: profile.totals.max_active_groups,
-        entries,
-        slot_hits: profile.totals.slot_hits,
-        slot_misses: profile.totals.slot_misses,
-        evictions: profile.totals.evictions,
-        policies,
-    }
-}
-
-fn qwen36_route_profile_enabled() -> bool {
-    crate::prefill_ffi::metal_profile_enabled()
-        || std::env::var_os("SUPERSONIC_QWEN36_ROUTE_PROFILE").is_some()
-        || std::env::var_os("SUPERSONIC_QWEN36_ROUTE_PROFILE_DUMP_CALLS").is_some()
-        || std::env::var_os("SUPERSONIC_QWEN36_ROUTE_PROFILE_DUMP_TOPN_LAYERS").is_some()
-        || std::env::var_os("SUPERSONIC_QWEN36_MOE_BATCHED_PREFILL_FEASIBILITY").is_some()
-}
-
-pub fn qwen36_batched_prefill_feasibility_profile_enabled() -> bool {
-    crate::prefill_ffi::metal_profile_enabled()
-        || std::env::var_os("SUPERSONIC_QWEN36_MOE_BATCHED_PREFILL_FEASIBILITY").is_some()
-}
-
-pub fn qwen36_route_profile_record_active_experts(active_experts: &[usize]) {
-    if qwen36_route_profile_enabled() {
-        qwen36_route_profile_record(active_experts);
-    }
-}
-
-fn qwen36_batched_prefill_plan_chunk_sizes(current_chunk_size: usize) -> Vec<usize> {
-    let mut sizes: Vec<usize> =
-        if let Ok(raw) = std::env::var("SUPERSONIC_QWEN36_MOE_BATCHED_PREFILL_PLAN_CHUNKS") {
-            raw.split(',')
-                .filter_map(|part| part.trim().parse::<usize>().ok())
-                .filter(|&chunk| chunk > 0)
-                .collect()
-        } else {
-            QWEN36_BATCHED_PREFILL_PLAN_CHUNKS.to_vec()
-        };
-    sizes.push(current_chunk_size.max(1));
-    sizes
-}
-
-fn qwen36_expert_residency_profile_enabled() -> bool {
-    crate::prefill_ffi::metal_profile_enabled()
-        || std::env::var_os("SUPERSONIC_QWEN36_EXPERT_RESIDENCY_PROFILE").is_some()
-        || std::env::var_os("SUPERSONIC_QWEN36_PACK_CACHE_PROFILE").is_some()
-}
-
-#[allow(clippy::too_many_arguments)]
-fn qwen36_expert_residency_profile_record(
-    key: Qwen36ExpertResidencyProfileKey,
-    exact_hit: bool,
-    route_refill: bool,
-    allocation: bool,
-    active_groups: usize,
-    copied_bytes: usize,
-    slot_hits: usize,
-    slot_misses: usize,
-    evictions: usize,
-) {
-    if !qwen36_expert_residency_profile_enabled() {
-        return;
-    }
-    let profile = QWEN36_EXPERT_RESIDENCY_PROFILE
-        .get_or_init(|| Mutex::new(Qwen36ExpertResidencyProfileAccumulator::default()));
-    let mut profile = profile
-        .lock()
-        .expect("qwen36 expert residency profile mutex poisoned");
-    qwen36_expert_residency_counters_record(
-        &mut profile.totals,
-        exact_hit,
-        route_refill,
-        allocation,
-        active_groups,
-        copied_bytes,
-        slot_hits,
-        slot_misses,
-        evictions,
-    );
-    let policy = profile.policies.entry(key).or_default();
-    qwen36_expert_residency_counters_record(
-        policy,
-        exact_hit,
-        route_refill,
-        allocation,
-        active_groups,
-        copied_bytes,
-        slot_hits,
-        slot_misses,
-        evictions,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn qwen36_packed_expert_cache_profile_record(
-    exact_hit: bool,
-    route_refill: bool,
-    allocation: bool,
-    active_groups: usize,
-    copied_bytes: usize,
-    slot_hits: usize,
-    slot_misses: usize,
-    evictions: usize,
-) {
-    qwen36_expert_residency_profile_record(
-        Qwen36ExpertResidencyProfileKey::new(
-            Qwen36ExpertResidencyMissPolicy::ExactRoute,
-            active_groups,
-        ),
-        exact_hit,
-        route_refill,
-        allocation,
-        active_groups,
-        copied_bytes,
-        slot_hits,
-        slot_misses,
-        evictions,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn qwen36_hotset_expert_residency_profile_record(
-    capacity: usize,
-    exact_hit: bool,
-    route_refill: bool,
-    allocation: bool,
-    active_groups: usize,
-    copied_bytes: usize,
-    slot_hits: usize,
-    slot_misses: usize,
-    evictions: usize,
-) {
-    qwen36_expert_residency_profile_record(
-        Qwen36ExpertResidencyProfileKey::new(Qwen36ExpertResidencyMissPolicy::LruHotset, capacity),
-        exact_hit,
-        route_refill,
-        allocation,
-        active_groups,
-        copied_bytes,
-        slot_hits,
-        slot_misses,
-        evictions,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn qwen36_static_topn_expert_residency_profile_record(
-    capacity: usize,
-    exact_hit: bool,
-    route_refill: bool,
-    allocation: bool,
-    active_groups: usize,
-    copied_bytes: usize,
-    slot_hits: usize,
-    slot_misses: usize,
-    evictions: usize,
-) {
-    qwen36_expert_residency_profile_record(
-        Qwen36ExpertResidencyProfileKey::new(Qwen36ExpertResidencyMissPolicy::StaticTopN, capacity),
-        exact_hit,
-        route_refill,
-        allocation,
-        active_groups,
-        copied_bytes,
-        slot_hits,
-        slot_misses,
-        evictions,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn qwen36_mps_static_topn_expert_residency_profile_record(
-    capacity: usize,
-    exact_hit: bool,
-    route_refill: bool,
-    allocation: bool,
-    active_groups: usize,
-    copied_bytes: usize,
-    slot_hits: usize,
-    slot_misses: usize,
-    evictions: usize,
-) {
-    qwen36_expert_residency_profile_record(
-        Qwen36ExpertResidencyProfileKey::with_format(
-            Qwen36ExpertResidentFormat::Fp16Mps,
-            Qwen36ExpertResidencyMissPolicy::StaticTopN,
-            capacity,
-        ),
-        exact_hit,
-        route_refill,
-        allocation,
-        active_groups,
-        copied_bytes,
-        slot_hits,
-        slot_misses,
-        evictions,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn qwen36_gpu_pack_expert_residency_profile_record(
-    exact_hit: bool,
-    route_refill: bool,
-    allocation: bool,
-    active_groups: usize,
-    copied_bytes: usize,
-    slot_hits: usize,
-    slot_misses: usize,
-    evictions: usize,
-) {
-    qwen36_expert_residency_profile_record(
-        Qwen36ExpertResidencyProfileKey::new(
-            Qwen36ExpertResidencyMissPolicy::GpuPack,
-            active_groups,
-        ),
-        exact_hit,
-        route_refill,
-        allocation,
-        active_groups,
-        copied_bytes,
-        slot_hits,
-        slot_misses,
-        evictions,
-    );
-}
-
-fn qwen36_route_profile_layers() -> usize {
-    std::env::var("SUPERSONIC_QWEN36_ROUTE_PROFILE_LAYERS")
-        .ok()
-        .and_then(|value| value.parse::<usize>().ok())
-        .filter(|&value| value > 0)
-        .unwrap_or(QWEN36_ROUTE_PROFILE_DEFAULT_LAYERS)
-}
-
-fn qwen36_route_profile_capacities() -> Vec<usize> {
-    let mut capacities = std::env::var("SUPERSONIC_QWEN36_ROUTE_PROFILE_CAPACITIES")
-        .ok()
-        .map(|value| {
-            value
-                .split(',')
-                .filter_map(|part| part.trim().parse::<usize>().ok())
-                .filter(|&capacity| capacity > 0)
-                .collect::<Vec<_>>()
-        })
-        .filter(|capacities| !capacities.is_empty())
-        .unwrap_or_else(|| QWEN36_ROUTE_PROFILE_DEFAULT_CAPS.to_vec());
-    capacities.sort_unstable();
-    capacities.dedup();
-    capacities
-}
-
-fn qwen36_route_profile_max_calls() -> usize {
-    std::env::var("SUPERSONIC_QWEN36_ROUTE_PROFILE_MAX_CALLS")
-        .ok()
-        .and_then(|value| value.parse::<usize>().ok())
-        .filter(|&value| value > 0)
-        .unwrap_or(QWEN36_ROUTE_PROFILE_DEFAULT_MAX_CALLS)
-}
-
-fn qwen36_route_profile_record(active_experts: &[usize]) {
-    let profile =
-        QWEN36_ROUTE_PROFILE.get_or_init(|| Mutex::new(Qwen36RouteProfileAccumulator::default()));
-    let mut profile = profile.lock().expect("qwen36 route profile mutex poisoned");
-    if profile.records.len() >= qwen36_route_profile_max_calls() {
-        profile.dropped_calls += 1;
-        return;
-    }
-    profile.records.push(
-        active_experts
-            .iter()
-            .map(|&expert| expert.min(u16::MAX as usize) as u16)
-            .collect(),
-    );
-}
-
-fn qwen36_route_profile_simulate(
-    records: &[Vec<u16>],
-    dropped_calls: u64,
-    layers: usize,
-) -> Qwen36RouteProfileSnapshot {
-    let mut snapshot = Qwen36RouteProfileSnapshot {
-        calls: records.len() as u64,
-        dropped_calls,
-        layers,
-        ..Qwen36RouteProfileSnapshot::default()
-    };
-    if records.is_empty() || layers == 0 {
-        return snapshot;
-    }
-
-    let mut unique_layer_experts = HashMap::<(usize, u16), ()>::new();
-    let mut per_layer_freq: Vec<HashMap<u16, u64>> = (0..layers).map(|_| HashMap::new()).collect();
-    for (call_idx, experts) in records.iter().enumerate() {
-        let layer = call_idx % layers;
-        snapshot.route_calls.push(Qwen36RouteCall {
-            call_idx,
-            layer,
-            experts: experts.clone(),
-        });
-        snapshot.assignments += experts.len() as u64;
-        for &expert in experts {
-            unique_layer_experts.insert((layer, expert), ());
-            *per_layer_freq[layer].entry(expert).or_insert(0) += 1;
-        }
-        if call_idx >= layers {
-            let prev = &records[call_idx - layers];
-            for &expert in experts {
-                snapshot.adjacent_total += 1;
-                if prev.contains(&expert) {
-                    snapshot.adjacent_hits += 1;
-                }
-            }
-        }
-    }
-    snapshot.unique_layer_experts = unique_layer_experts.len();
-
-    for capacity in qwen36_route_profile_capacities() {
-        let mut caches: Vec<Vec<u16>> = (0..layers).map(|_| Vec::new()).collect();
-        let mut sim = Qwen36RouteCacheSim {
-            capacity,
-            ..Qwen36RouteCacheSim::default()
-        };
-        for (call_idx, experts) in records.iter().enumerate() {
-            let layer = call_idx % layers;
-            let cache = &mut caches[layer];
-            for &expert in experts {
-                if let Some(pos) = cache.iter().position(|&cached| cached == expert) {
-                    sim.hits += 1;
-                    let expert = cache.remove(pos);
-                    cache.push(expert);
-                } else {
-                    sim.misses += 1;
-                    if cache.len() >= capacity {
-                        cache.remove(0);
-                    }
-                    cache.push(expert);
-                }
-            }
-        }
-        snapshot.cache_sims.push(sim);
-
-        let mut topn = Qwen36RouteTopNSim {
-            capacity,
-            total: snapshot.assignments,
-            ..Qwen36RouteTopNSim::default()
-        };
-        for (layer, freq) in per_layer_freq.iter().enumerate() {
-            let mut counts: Vec<(u16, u64)> = freq
-                .iter()
-                .map(|(&expert, &count)| (expert, count))
-                .collect();
-            counts.sort_unstable_by(|(lhs_expert, lhs_count), (rhs_expert, rhs_count)| {
-                rhs_count
-                    .cmp(lhs_count)
-                    .then_with(|| lhs_expert.cmp(rhs_expert))
-            });
-            let selected: Vec<(u16, u64)> = counts.into_iter().take(capacity).collect();
-            let covered = selected.iter().map(|(_, count)| *count).sum::<u64>();
-            let total = freq.values().copied().sum::<u64>();
-            topn.covered += covered;
-            snapshot.topn_layers.push(Qwen36RouteTopNLayer {
-                capacity,
-                layer,
-                experts: selected.iter().map(|(expert, _)| *expert).collect(),
-                counts: selected.iter().map(|(_, count)| *count).collect(),
-                covered,
-                total,
-            });
-        }
-        snapshot.topn_sims.push(topn);
-    }
-
-    snapshot
-}
-
-fn qwen36_batched_prefill_feasibility_profile_simulate(
-    records: &[Vec<u16>],
-    dropped_calls: u64,
-    config: &Qwen36BatchedPrefillFeasibilityConfig,
-) -> Qwen36BatchedPrefillFeasibilityProfileSnapshot {
-    let layers = config.layers.max(1);
-    let top_k = config.top_k.max(1);
-    let chunk_size = config.chunk_size.max(1);
-    let available_tokens = records.len() / layers;
-    let requested_tokens = if config.prefill_tokens == 0 {
-        available_tokens
-    } else {
-        config.prefill_tokens
-    };
-    let profiled_tokens = requested_tokens.min(available_tokens);
-    let chunks = if profiled_tokens == 0 {
-        0
-    } else {
-        profiled_tokens.div_ceil(chunk_size)
-    };
-    let mut snapshot = Qwen36BatchedPrefillFeasibilityProfileSnapshot {
-        calls: records.len() as u64,
-        dropped_calls,
-        layers,
-        top_k,
-        num_experts: config.num_experts.max(1),
-        chunk_size,
-        prefill_tokens: config.prefill_tokens,
-        profiled_tokens,
-        chunks,
-        ..Qwen36BatchedPrefillFeasibilityProfileSnapshot::default()
-    };
-    if profiled_tokens == 0 {
-        return snapshot;
-    }
-
-    for chunk_start in (0..profiled_tokens).step_by(chunk_size) {
-        let chunk_end = (chunk_start + chunk_size).min(profiled_tokens);
-        for layer in 0..layers {
-            let mut counts = HashMap::<u16, u64>::new();
-            for token in chunk_start..chunk_end {
-                let call_idx = token * layers + layer;
-                let Some(experts) = records.get(call_idx) else {
-                    continue;
-                };
-                for &expert in experts.iter().take(top_k) {
-                    *counts.entry(expert).or_insert(0) += 1;
-                    snapshot.assignments += 1;
-                }
-            }
-            snapshot.expert_segments += counts.len() as u64;
-            snapshot.max_unique_experts_per_layer_chunk = snapshot
-                .max_unique_experts_per_layer_chunk
-                .max(counts.len());
-            for rows in counts.values().copied() {
-                snapshot.max_rows_per_segment = snapshot.max_rows_per_segment.max(rows);
-                snapshot.wmma16_padded_assignments += rows.div_ceil(16) * 16;
-                if rows >= 16 {
-                    snapshot.wmma16_segments += 1;
-                    snapshot.wmma16_covered_assignments += rows;
-                } else {
-                    snapshot.scalar_tail_segments += 1;
-                    snapshot.scalar_tail_assignments += rows;
-                }
-            }
-        }
-    }
-    snapshot.permutation_entries = snapshot.assignments;
-    snapshot
-}
-
-/// Per-layer descriptor for the Qwen3.6-MoE megakernel. Field order and
-/// natural x86_64 alignment must match the C++ struct in
-/// `kernels/qwen36_moe.hip` exactly. The repr-C layout is fixed at PR 4
-/// time and grows by appending new fields, never reordering existing
-/// ones — see the matching `static_assert(sizeof(...))` on the C++ side.
-#[repr(C)]
-#[derive(Debug, Clone)]
-pub struct Qwen36MoeDecodeLayerDesc {
-    /// Layer index in `[0, num_hidden_layers)`. Used by the kernel to pick
-    /// the cos/sin RoPE entry and to sanity-check the descriptor pointer.
-    pub layer_idx: c_int,
-    /// 0 = linear-attention layer, 1 = full-attention layer.
-    pub is_full_attention: c_int,
-
-    // --- RMS norms --------------------------------------------------------
-    pub input_norm_w: *const c_void,
-    pub input_norm_eps: f32,
-    pub post_attn_norm_w: *const c_void,
-    pub post_attn_norm_eps: f32,
-
-    // --- Full-attention slots (read iff is_full_attention == 1) -----------
-    /// q_proj output dim. With `attn_output_gate=true` (Qwen3-Next) this is
-    /// `2 * num_heads * head_dim`; the kernel splits the upper half off as
-    /// the sigmoid output gate. With `attn_output_gate=false` it's just
-    /// `num_heads * head_dim`. The sign is captured by `attn_output_gate`.
-    pub q_proj_w: *const c_void,
-    pub q_proj_out_dim: c_int,
-    /// 0 = no output gate (q_proj_out_dim == num_heads*head_dim),
-    /// 1 = attn_output_gate fused (q_proj_out_dim == 2*num_heads*head_dim).
-    pub attn_output_gate: c_int,
-    pub k_proj_w: *const c_void,
-    pub v_proj_w: *const c_void,
-    pub o_proj_w: *const c_void,
-    pub q_norm_w: *const c_void,
-    pub k_norm_w: *const c_void,
-    pub attn_head_dim: c_int,
-    pub attn_num_heads: c_int,
-    pub attn_num_kv_heads: c_int,
-    pub kv_cache_k: *mut c_void,
-    pub kv_cache_v: *mut c_void,
-    pub kv_len: c_int,
-    pub kv_max_t: c_int,
-
-    // --- Linear-attention slots (read iff is_full_attention == 0) ---------
-    pub linear_in_proj_qkv_w: *const c_void,
-    pub linear_in_proj_z_w: *const c_void,
-    pub linear_in_proj_b_w: *const c_void,
-    pub linear_in_proj_a_w: *const c_void,
-    pub linear_out_proj_w: *const c_void,
-    pub linear_conv1d_w: *const c_void,
-    pub linear_dt_bias: *const c_void,
-    pub linear_a_log_exp: *const c_void,
-    pub linear_norm_w: *const c_void,
-    pub linear_qkv_dim: c_int,
-    pub linear_v_dim: c_int,
-    pub linear_v_heads: c_int,
-    pub linear_conv_kernel_dim: c_int,
-    /// Linear-attention conv state pointer, shape `[batch, qkv_dim,
-    /// kernel-1]`. NULL on first decode step (kernel will zero on read).
-    pub linear_conv_state: *mut c_void,
-    /// Linear-attention recurrent state, shape `[batch, V_heads, V_dim,
-    /// K_dim]`. NULL on first decode step.
-    pub linear_recurrent_state: *mut c_void,
-
-    // --- MoE block (always read, regardless of attention type) ------------
-    /// Router weight `[num_experts, hidden]`, BF16. Always BF16 (excluded
-    /// from INT4 quant by `is_int4_target`).
-    pub router_w: *const c_void,
-    /// Fused expert gate+up `[num_experts, 2*moe_intermediate_size, hidden]`.
-    /// At INT4 launch the pointer reinterprets as packed `u8` (2 nibbles
-    /// per byte), with sidecar scale/zero in `Qwen36MoeInt4ScaleDesc`.
-    pub experts_gate_up_w: *const c_void,
-    /// Fused expert down `[num_experts, hidden, moe_intermediate_size]`.
-    pub experts_down_w: *const c_void,
-    /// Shared expert (always-on). `gate_proj` and `up_proj` are
-    /// `[shared_int, hidden]`; `down_proj` is `[hidden, shared_int]`.
-    pub shared_expert_gate_proj_w: *const c_void,
-    pub shared_expert_up_proj_w: *const c_void,
-    pub shared_expert_down_proj_w: *const c_void,
-    /// Scalar shared-expert gate `[1, hidden]`, BF16. Applied as
-    /// `sigmoid(gate · x) * shared_expert(x)`.
-    pub shared_expert_gate_w: *const c_void,
-    /// Number of routed experts present in this layer. Must match
-    /// `desc.num_experts` across layers (sanity-checked by the host).
-    pub num_experts: c_int,
-    /// Top-k for routing.
-    pub top_k: c_int,
-    pub moe_intermediate_size: c_int,
-    pub shared_expert_intermediate_size: c_int,
-    /// 1 if router applies `softmax(top_k_logits)` renormalization
-    /// (`norm_topk_prob=true` in config). 0 otherwise.
-    pub norm_topk_prob: c_int,
-
-    // --- KV-FP8 sidecar (read iff is_full_attention == 1 AND
-    // matching kv_fp8_descs[layer].kv_scale_k != null) ---------------
-    /// BF16 sidecar buffer `[num_kv_heads, kv_shadow_window, head_dim]`.
-    /// Null when the sidecar is disabled. The kernel reads from the sidecar
-    /// (instead of dequantising FP8) for positions covered by the rolling
-    /// sidecar window.
-    pub kv_shadow_k: *mut c_void,
-    /// BF16 sidecar buffer `[num_kv_heads, kv_shadow_window, head_dim]`.
-    /// Paired with [`Self::kv_shadow_k`]; null under the same conditions.
-    pub kv_shadow_v: *mut c_void,
-    /// Earliest absolute KV position the sidecar may cover. `-1` when the
-    /// sidecar is disabled. Runtime coverage is
-    /// `max(kv_shadow_start, position + 1 - kv_shadow_window)..=position`.
-    pub kv_shadow_start: c_int,
-    /// Number of recent KV positions physically stored in the BF16 sidecar.
-    /// Zero when the sidecar is disabled. The kernel uses modulo indexing so
-    /// the descriptor can remain fixed across decode steps.
-    pub kv_shadow_window: c_int,
-}
-
-unsafe impl Send for Qwen36MoeDecodeLayerDesc {}
-unsafe impl Sync for Qwen36MoeDecodeLayerDesc {}
-
-impl Default for Qwen36MoeDecodeLayerDesc {
-    fn default() -> Self {
-        unsafe { std::mem::zeroed() }
-    }
-}
-
-/// Parallel-struct to [`Qwen36MoeDecodeLayerDesc`] carrying INT4 GPTQ
-/// scale + zero pointers. When `--int4` is active, the main desc's `*_w`
-/// slots reinterpret as packed-u8 nibbles and the kernel reads the
-/// matching `*_scale` / `*_zero` from this struct.
-///
-/// Routers and scalar gates stay BF16 (`is_int4_target` excludes them);
-/// their fields here are unused but the struct keeps them at fixed offsets
-/// for ABI stability.
-#[repr(C)]
-#[derive(Debug, Clone)]
-pub struct Qwen36MoeInt4ScaleDesc {
-    pub q_proj_scale: *const c_void,
-    pub q_proj_zero: *const c_void,
-    pub k_proj_scale: *const c_void,
-    pub k_proj_zero: *const c_void,
-    pub v_proj_scale: *const c_void,
-    pub v_proj_zero: *const c_void,
-    pub o_proj_scale: *const c_void,
-    pub o_proj_zero: *const c_void,
-
-    pub linear_in_proj_qkv_scale: *const c_void,
-    pub linear_in_proj_qkv_zero: *const c_void,
-    pub linear_in_proj_z_scale: *const c_void,
-    pub linear_in_proj_z_zero: *const c_void,
-    pub linear_out_proj_scale: *const c_void,
-    pub linear_out_proj_zero: *const c_void,
-
-    pub experts_gate_up_scale: *const c_void,
-    pub experts_gate_up_zero: *const c_void,
-    pub experts_down_scale: *const c_void,
-    pub experts_down_zero: *const c_void,
-
-    pub shared_expert_gate_proj_scale: *const c_void,
-    pub shared_expert_gate_proj_zero: *const c_void,
-    pub shared_expert_up_proj_scale: *const c_void,
-    pub shared_expert_up_proj_zero: *const c_void,
-    pub shared_expert_down_proj_scale: *const c_void,
-    pub shared_expert_down_proj_zero: *const c_void,
-
-    pub group_size: c_int,
-}
-
-unsafe impl Send for Qwen36MoeInt4ScaleDesc {}
-unsafe impl Sync for Qwen36MoeInt4ScaleDesc {}
-
-impl Default for Qwen36MoeInt4ScaleDesc {
-    fn default() -> Self {
-        unsafe { std::mem::zeroed() }
-    }
-}
-
-/// Per-layer KV cache FP8 scale pointers for Qwen3.6-MoE.
-///
-/// Parallel struct to [`Qwen36MoeDecodeLayerDesc`] — one entry per layer,
-/// passed as a separate kernel argument (same pattern as
-/// [`Qwen36MoeInt4ScaleDesc`]). Linear-attention layers leave both
-/// pointers null. When KV-FP8 is off, the entire
-/// `*const Qwen36MoeKVCacheFp8Desc` array argument is null.
-///
-/// Mirrors the qwen35 `KVCacheFp8Desc` shape: F32 absmax scale per
-/// (kv_head, position).
-#[repr(C)]
-#[derive(Debug, Clone)]
-pub struct Qwen36MoeKVCacheFp8Desc {
-    /// `[num_kv_heads, max_T]` F32. Null for linear-attn layers.
-    pub kv_scale_k: *mut c_void,
-    /// `[num_kv_heads, max_T]` F32. Null for linear-attn layers.
-    pub kv_scale_v: *mut c_void,
-}
-
-unsafe impl Send for Qwen36MoeKVCacheFp8Desc {}
-unsafe impl Sync for Qwen36MoeKVCacheFp8Desc {}
-
-impl Default for Qwen36MoeKVCacheFp8Desc {
-    fn default() -> Self {
-        unsafe { std::mem::zeroed() }
-    }
-}
-
-/// Per-sequence batched-decode state, parallel to the layer descriptor
-/// array. Only the first `batch_size` slots are read.
-#[repr(C)]
-#[derive(Debug, Clone)]
-pub struct Qwen36MoeBatchSeqDesc {
-    pub seqlen_offset: [c_int; MAX_BATCH_SIZE],
-    pub kv_cache_k: [*mut c_void; MAX_BATCH_SIZE],
-    pub kv_cache_v: [*mut c_void; MAX_BATCH_SIZE],
-    pub kv_len: [c_int; MAX_BATCH_SIZE],
-    pub kv_max_t: [c_int; MAX_BATCH_SIZE],
-    pub linear_conv_state: [*mut c_void; MAX_BATCH_SIZE],
-    pub linear_recurrent_state: [*mut c_void; MAX_BATCH_SIZE],
-}
-
-unsafe impl Send for Qwen36MoeBatchSeqDesc {}
-unsafe impl Sync for Qwen36MoeBatchSeqDesc {}
-
-impl Default for Qwen36MoeBatchSeqDesc {
-    fn default() -> Self {
-        unsafe { std::mem::zeroed() }
-    }
-}
-
-/// Attribution-only MPP pilot used by the Apple M5 Metal bench harness.
-///
-/// This measures repeated exact `64x32x64` MPP tensor tiles as an equivalent
-/// square GEMM throughput number. It does not consume Qwen3.6 model weights
-/// and must not be interpreted as a decode-path replacement.
-pub fn metal_mpp_tile_gemm_f16_tflops(size: u32, iterations: u32) -> Result<f64, GpuError> {
-    crate::metal_native::mpp_tile_gemm_f16_tflops(size, iterations)
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct MetalMpsExpertF16Probe {
-    pub gate_up_ms: f64,
-    pub down_ms: f64,
-    pub gate_up_tflops: f64,
-    pub down_tflops: f64,
-}
-
-/// Attribution-only MPS probe for Qwen3.6 active-expert GEMV shapes.
-///
-/// This is a resident-FP16 vendor-library upper-bound probe. It does not use the
-/// GPTQ INT4 expert buffers directly and does not change the decode path.
-pub fn metal_mps_expert_f16_probe(
-    hidden: usize,
-    moe_intermediate: usize,
-    top_k: usize,
-    iterations: u32,
-) -> Result<MetalMpsExpertF16Probe, GpuError> {
-    let probe = crate::metal_native::qwen36_mps_expert_f16_probe(
-        hidden,
-        moe_intermediate,
-        top_k,
-        iterations,
-    )?;
-    Ok(MetalMpsExpertF16Probe {
-        gate_up_ms: probe.gate_up_ms,
-        down_ms: probe.down_ms,
-        gate_up_tflops: probe.gate_up_tflops,
-        down_tflops: probe.down_tflops,
-    })
-}
-
-#[cfg(any(supersonic_backend_hip, supersonic_backend_cuda))]
-extern "C" {
-    /// Stub launch entry. Walks the descriptor array, validates field
-    /// integrity by writing recognizable sentinel values into the workspace
-    /// at known offsets, grid-barriers between layers, and returns 0 on
-    /// success.
-    ///
-    /// Sentinel layout in `workspace[0..sentinel_count]` (f32):
-    /// - `[0]`: number of layers seen (must equal `num_layers`)
-    /// - `[1]`: total `num_experts` summed across layers (sanity check)
-    /// - `[2]`: total `top_k` summed across layers
-    /// - `[3]`: 1.0 if every layer's `is_full_attention` matches the
-    ///   pattern produced by `(idx + 1) % 4 == 0`, else 0.0
-    /// - `[4]`: `attn_output_gate` status — 1.0 if all full-attn layers
-    ///   set it to 1, 0.0 otherwise
-    /// - `[5..]`: reserved for future smoke-test bytes; zero on PR 4.
-    ///
-    /// Once the real kernel lands, this entry is replaced by the actual
-    /// persistent decode launcher with the same signature.
-    pub fn qwen36_moe_hip_stub_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        num_layers: usize,
-        layers: *const Qwen36MoeDecodeLayerDesc,
-        workspace: *mut f32,
-        counters: *mut c_uint,
-        barrier_counter: *mut c_uint,
-        barrier_flag: *mut c_uint,
-    ) -> c_int;
-
-    /// Phase 3e: persistent decode megakernel launcher. One cooperative
-    /// HIP launch processes all `num_layers` of {attn or linear-attn, FFN}
-    /// — replaces 80 step launches/token with 1 (the lm_head still
-    /// launches separately at this stage).
-    ///
-    /// See `kernels/qwen36_moe_persistent/persistent_decode.hip` for the
-    /// kernel and `kernels/qwen36_moe_bridge.cpp::qwen36_moe_hip_persistent_decode_launch`
-    /// for the launcher.
-    ///
-    /// Caller responsibilities:
-    /// - `hidden_ping` is uploaded with the initial hidden BF16 bytes; the
-    ///   final hidden lands back in `hidden_ping` after even `num_layers`
-    ///   (the bridge rejects odd `num_layers`).
-    /// - `int4_scales` is null for BF16 baked models, non-null for INT4
-    ///   bakes (one entry per layer, parallel to `layers`).
-    /// - `workspace` is at least
-    ///   `max(attn_workspace_floats(geom), ffn_workspace_floats(geom))` F32
-    ///   entries — same as the chained driver.
-    /// - `ffn_topk_idx_scratch` is a small `[top_k]` i32 buffer (the FFN
-    ///   phase only writes it at stage 1, but the parameter must be valid).
-    /// - sync_buf layout: counters[0..16] u32 + barrier_counter at +64 +
-    ///   barrier_flag at +68 (96 bytes total, zeroed by the bridge).
-    pub fn qwen36_moe_hip_persistent_decode_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        num_layers: c_int,
-        start_layer: c_int,
-        end_layer_exclusive: c_int,
-        mode: c_int,
-        layers: *const Qwen36MoeDecodeLayerDesc,
-        int4_scales: *const Qwen36MoeInt4ScaleDesc,
-        // Null when KV-FP8 is off globally. Otherwise an array of
-        // `num_layers` entries parallel to `layers`. Full-attention
-        // layers populate `kv_scale_k` / `kv_scale_v` together (both
-        // set, or both null to disable KV-FP8 for that layer specifically
-        // — a valid mixed-mode configuration). Linear-attention layers
-        // must leave both null. The bridge validates these invariants
-        // and rejects malformed descriptors before kernel launch.
-        kv_fp8_descs: *const Qwen36MoeKVCacheFp8Desc,
-        hidden: c_int,
-        num_heads: c_int,
-        num_kv_heads: c_int,
-        head_dim: c_int,
-        rotary_dim: c_int,
-        num_k_heads: c_int,
-        num_v_heads: c_int,
-        head_k_dim: c_int,
-        head_v_dim: c_int,
-        conv_kernel_dim: c_int,
-        num_experts: c_int,
-        moe_intermediate: c_int,
-        shared_intermediate: c_int,
-        top_k: c_int,
-        vocab: c_int,
-        rope_theta: f32,
-        rms_norm_eps: f32,
-        position: c_int,
-        // -1 ⇒ inherit from `position` (dense base-decode case);
-        // ≥ 0 ⇒ decoupled KV slot for SpecPrefill sparse-prefill or
-        // MTP draft layers.
-        cache_pos: c_int,
-        embed_w: *const c_void,
-        token_id: c_int,
-        token_ids: *const c_uint,
-        prefill_len: c_int,
-        hidden_ping: *mut c_void,
-        hidden_pong: *mut c_void,
-        workspace: *mut f32,
-        ffn_topk_idx_scratch: *mut c_int,
-        // Phase 3f folded final RMSnorm + lm_head GEMV. Pass nullptr
-        // triple + vocab=0 to skip (prefill steps); otherwise the
-        // megakernel writes logits to `logits_out` and the host can
-        // skip the separate `lm_head_launch` call.
-        final_norm_w: *const c_void,
-        lm_head_w: *const c_void,
-        logits_out: *mut c_void,
-        top1_out: *mut c_uint,
-        counters: *mut c_uint,
-        barrier_counter: *mut c_uint,
-        barrier_flag: *mut c_uint,
-    ) -> c_int;
-
-    /// PR 4b2 staged single-layer attention parity launcher. Runs the
-    /// full-attention path through `stage` (1..=5) and writes the matching
-    /// intermediate to `output`:
-    ///
-    /// | stage | output buffer contents (BF16)                      |
-    /// |-------|----------------------------------------------------|
-    /// |   1   | `q_normed[H*d]`                                    |
-    /// |   2   | `k_normed[Hkv*d]`         (`q_normed` recomputed)  |
-    /// |   3   | `q_rot[H*d] || k_rot[Hkv*d]` (planned)             |
-    /// |   4   | `attn[H*d]`                                        |
-    /// |   5   | `output_hidden[hidden]`                            |
-    ///
-    /// At PR 4b2 step 1 only `stage == 1` is wired; the kernel returns the
-    /// q-path intermediate and ignores the k_*/v_*/o_proj/RoPE/position
-    /// arguments. They're declared up front so the FFI ABI doesn't change
-    /// between staged commits.
-    ///
-    /// `workspace` must be at least `2 * num_heads * head_dim` F32 entries
-    /// (used to hold the BF16-rounded F32 view of `q_raw` between phases).
-    /// `output` must be at least `num_heads * head_dim` BF16 entries on
-    /// stage 1 — sized for the largest staged intermediate, BF16.
-    /// `sync_buf` (counters/barrier_counter/barrier_flag) must be 96 zero
-    /// bytes — see [`stub_launch`] for the layout convention.
-    pub fn qwen36_moe_hip_attn_step_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        stage: c_int,
-        hidden: c_int,
-        num_heads: c_int,
-        num_kv_heads: c_int,
-        head_dim: c_int,
-        rotary_dim: c_int,
-        rope_theta: f32,
-        rms_norm_eps: f32,
-        position: c_int,
-        cache_pos: c_int,
-        input_hidden: *const c_void,
-        input_norm_w: *const c_void,
-        q_proj_w: *const c_void,
-        k_proj_w: *const c_void,
-        v_proj_w: *const c_void,
-        q_norm_w: *const c_void,
-        k_norm_w: *const c_void,
-        o_proj_w: *const c_void,
-        int4_group_size: c_int,
-        q_proj_scale: *const c_void,
-        q_proj_zero: *const c_void,
-        k_proj_scale: *const c_void,
-        k_proj_zero: *const c_void,
-        v_proj_scale: *const c_void,
-        v_proj_zero: *const c_void,
-        o_proj_scale: *const c_void,
-        o_proj_zero: *const c_void,
-        output: *mut c_void,
-        workspace: *mut f32,
-        kv_cache_k: *mut c_void,
-        kv_cache_v: *mut c_void,
-        kv_max_t: c_int,
-        counters: *mut c_uint,
-        barrier_counter: *mut c_uint,
-        barrier_flag: *mut c_uint,
-    ) -> c_int;
-
-    /// PR 4b3 staged single-layer linear-attention parity launcher. Same
-    /// staged-build-up discipline as `qwen36_moe_hip_attn_step_launch`,
-    /// but for the 3-of-4 hybrid layers that aren't full-attention.
-    /// `stage` selects how far to run; the matching staged intermediate
-    /// is published to `output` (BF16):
-    ///
-    /// | stage | output buffer contents (BF16)           |
-    /// |-------|------------------------------------------|
-    /// |   1   | `qkv_raw[qkv_dim]`                       |
-    /// |   2   | `silu_out[qkv_dim]`         (planned)    |
-    /// |   3   | `q_scaled || k_rep || v_heads` (planned) |
-    /// |   4   | `recurrent_out[V*v_dim]`    (planned)    |
-    /// |   5   | `output_hidden[hidden]`     (planned)    |
-    ///
-    /// PR 4b3 step 2 wires only `stage == 1`; the kernel ignores the conv
-    /// / dt / norm / out_proj / state pointers and the matching arguments
-    /// can be null. They're declared up front so subsequent staged commits
-    /// don't perturb the FFI ABI.
-    ///
-    /// `workspace` must be at least `qkv_dim + V*v_dim + 2*V` F32 entries
-    /// for stage 1 (later stages bump that up via the safe wrapper).
-    /// `output` must be at least `qkv_dim` BF16 entries on stage 1 (sized
-    /// for the largest staged intermediate by the safe wrapper). `sync_buf`
-    /// (counters/barrier_counter/barrier_flag) must be 96 zero bytes.
-    pub fn qwen36_moe_hip_linear_step_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        stage: c_int,
-        hidden: c_int,
-        num_k_heads: c_int,
-        num_v_heads: c_int,
-        head_k_dim: c_int,
-        head_v_dim: c_int,
-        conv_kernel_dim: c_int,
-        rms_norm_eps: f32,
-        input_hidden: *const c_void,
-        input_norm_w: *const c_void,
-        in_proj_qkv_w: *const c_void,
-        in_proj_z_w: *const c_void,
-        in_proj_a_w: *const c_void,
-        in_proj_b_w: *const c_void,
-        conv1d_w: *const c_void,
-        conv1d_bias: *const c_void,
-        dt_bias: *const c_void,
-        a_log: *const c_void,
-        norm_w: *const c_void,
-        out_proj_w: *const c_void,
-        conv_state: *mut c_void,
-        recurrent_state: *mut f32,
-        int4_group_size: c_int,
-        in_proj_qkv_scale: *const c_void,
-        in_proj_qkv_zero: *const c_void,
-        in_proj_z_scale: *const c_void,
-        in_proj_z_zero: *const c_void,
-        out_proj_scale: *const c_void,
-        out_proj_zero: *const c_void,
-        output: *mut c_void,
-        workspace: *mut f32,
-        counters: *mut c_uint,
-        barrier_counter: *mut c_uint,
-        barrier_flag: *mut c_uint,
-    ) -> c_int;
-
-    /// PR 4b4 staged single-block MoE FFN parity launcher. Same staged-build-up
-    /// discipline as `qwen36_moe_hip_attn_step_launch` and
-    /// `qwen36_moe_hip_linear_step_launch`, but for the post-attention half
-    /// of one Qwen3.6-MoE layer. `stage` selects how far to run; the matching
-    /// staged intermediate is published to `output` (BF16) and `output_idx`
-    /// (i32, top-k indices for stages 1+):
-    ///
-    /// | stage | output buffer contents (BF16)                    |
-    /// |-------|--------------------------------------------------|
-    /// |   1   | `topk_weights[k]`           (idx via `output_idx`) |
-    /// |   2   | `shared_out[hidden]`                             |
-    /// |   3   | `expert_0_out[hidden]`      (top-1 dispatch)     |
-    /// |   4   | `moe_out[hidden]`                                |
-    /// |   5   | `output_hidden[hidden]`     (final residual)     |
-    ///
-    /// PR 4b4 step 1 wires only `stage == 1`; the kernel ignores the
-    /// gate_up_proj / down_proj / shared_expert_* pointers and the matching
-    /// arguments can be null. They're declared up front so subsequent staged
-    /// commits don't perturb the FFI ABI.
-    ///
-    /// `workspace` must be at least `hidden + 2*num_experts + 2*top_k` F32
-    /// entries for stage 1 (later stages bump that up). `output` must be at
-    /// least `top_k` BF16 entries on stage 1 and `output_idx` must be at
-    /// least `top_k` i32 entries. `sync_buf` (counters/barrier_counter/
-    /// barrier_flag) must be 96 zero bytes.
-    ///
-    /// PR 4b5 step 2: INT4 dequant smoke launcher.
-    ///
-    /// Drives a tiny single-thread kernel that runs both `int4_dequant_8`
-    /// and `int4_dequant_scalar` over a `[out_rows, in_cols]` slab, writing
-    /// each helper's outputs into a separate buffer. The Rust-side test
-    /// validates byte-for-byte against a host reference computing the same
-    /// `bf16(q*s - z*s)` reconstruction. Catches porting bugs in the
-    /// helpers in isolation, before they're folded into the real FFN
-    /// matmuls in step 3+.
-    ///
-    /// `packed`: u8, shape `[out_rows, in_cols / 2]`, even col → low nibble.
-    /// `scale` / `zero`: BF16, shape `[out_rows / gsz, in_cols / gsz]`.
-    /// `dq_8_out`, `dq_scalar_out`: F32 device buffers, each
-    /// `out_rows * in_cols` long.
-    ///
-    /// Pre-conditions (the bridge validates them):
-    /// - `in_cols % 8 == 0`
-    /// - `in_cols % gsz == 0` and `gsz % 2 == 0`
-    /// - `out_rows % gsz == 0`
-    pub fn qwen36_moe_hip_int4_dequant_smoke_launch(
-        device_ordinal: usize,
-        packed: *const u8,
-        scale: *const c_void,
-        zero: *const c_void,
-        out_rows: c_int,
-        in_cols: c_int,
-        gsz: c_int,
-        dq_8_out: *mut f32,
-        dq_scalar_out: *mut f32,
-    ) -> c_int;
-
-    pub fn qwen36_moe_hip_ffn_step_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        stage: c_int,
-        hidden: c_int,
-        num_experts: c_int,
-        moe_intermediate: c_int,
-        shared_intermediate: c_int,
-        top_k: c_int,
-        rms_norm_eps: f32,
-        input_hidden: *const c_void,
-        post_attn_norm_w: *const c_void,
-        gate_w: *const c_void,
-        gate_up_proj_w: *const c_void,
-        down_proj_w: *const c_void,
-        shared_gate_proj_w: *const c_void,
-        shared_up_proj_w: *const c_void,
-        shared_down_proj_w: *const c_void,
-        shared_expert_gate_w: *const c_void,
-        int4_group_size: c_int,
-        gate_up_proj_scale: *const c_void,
-        gate_up_proj_zero: *const c_void,
-        down_proj_scale: *const c_void,
-        down_proj_zero: *const c_void,
-        shared_gate_proj_scale: *const c_void,
-        shared_gate_proj_zero: *const c_void,
-        shared_up_proj_scale: *const c_void,
-        shared_up_proj_zero: *const c_void,
-        shared_down_proj_scale: *const c_void,
-        shared_down_proj_zero: *const c_void,
-        output: *mut c_void,
-        output_idx: *mut c_int,
-        workspace: *mut f32,
-        counters: *mut c_uint,
-        barrier_counter: *mut c_uint,
-        barrier_flag: *mut c_uint,
-    ) -> c_int;
-
-    /// Final RMSNorm + lm_head GEMV in a single kernel — replaces the
-    /// host-side `host_final_norm_lm_head_f32` for qwen3.6-MoE.
-    ///
-    /// All buffers are device pointers, all BF16 (`dtype = 2`):
-    ///   - `final_hidden`: [hidden] BF16, the output of `run_chained_decode`.
-    ///   - `final_norm_w`: [hidden] BF16 — `model.norm.weight`. Applies the
-    ///      HF `Qwen3_5MoeRMSNorm` `(1 + w)` unit offset.
-    ///   - `lm_head_w`: [vocab, hidden] BF16, dequantized once at startup.
-    ///   - `logits`: [vocab] BF16, output.
-    ///   - `counter`: [1] u32. Used as a work-stealing atomic across vocab
-    ///     rows; the launcher memsets it to 0 before each call so the
-    ///     caller doesn't need to.
-    ///
-    /// Returns 0 on success; non-zero on validation / launch failure (see
-    /// `qwen36_moe_hip_lm_head_launch` in `kernels/qwen36_moe_bridge.cpp`
-    /// for the error code matrix).
-    pub fn qwen36_moe_hip_lm_head_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        hidden: c_int,
-        vocab: c_int,
-        rms_norm_eps: f32,
-        final_hidden: *const c_void,
-        final_norm_w: *const c_void,
-        lm_head_w: *const c_void,
-        logits: *mut c_void,
-        // Optional BF16 [hidden] export of the post-RMSNorm hidden
-        // state. Phase 6.2c.3 plumbing for the MTP draft loop's
-        // recurrent feed; null = base-decode behavior unchanged.
-        x_normed_out: *mut c_void,
-        counter: *mut c_uint,
-    ) -> c_int;
-
-    /// FFI bridge for the batched lm_head WMMA kernel (Phase 6.4a). Wraps
-    /// `qwen36_moe_lm_head_batched_wmma_kernel`. `m` is the runtime batch
-    /// size (1..16); for `m == 1` the single-M path
-    /// (`qwen36_moe_hip_lm_head_launch`) is faster — use the batched
-    /// launcher when `m >= 2` to amortize the lm_head BF16 weight read.
-    /// WMMA-only (gfx11xx); returns status 138 on unsupported hardware
-    /// or `hidden % 16 != 0` so the caller can fall back to a per-row
-    /// loop over the single-M launcher.
-    pub fn qwen36_moe_hip_lm_head_batched_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        m: c_int,
-        hidden: c_int,
-        vocab: c_int,
-        rms_norm_eps: f32,
-        final_hidden: *const c_void,
-        final_norm_w: *const c_void,
-        lm_head_w: *const c_void,
-        logits: *mut c_void,
-        x_normed_out: *mut c_void,
-    ) -> c_int;
-
-    /// FFI bridge for the MTP pre-fusion kernel (Phase 6.2c.1). Single-block
-    /// launch: BF16 RMSNorms over `e_in` and `h_base` followed by a
-    /// `mtp.fc @ cat([e_norm, h_norm])` matvec into `fused_out`. All buffers
-    /// must be device-resident on `device_ordinal` and BF16. See
-    /// `qwen36_moe_hip_mtp_pre_fusion_launch` in `kernels/qwen36_moe_bridge.cpp`
-    /// for the error code matrix.
-    pub fn qwen36_moe_hip_mtp_pre_fusion_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        hidden: c_int,
-        rms_norm_eps: f32,
-        e_in: *const c_void,
-        h_base: *const c_void,
-        pre_fc_norm_embedding_w: *const c_void,
-        pre_fc_norm_hidden_w: *const c_void,
-        fc_w: *const c_void,
-        e_norm_out: *mut c_void,
-        h_norm_out: *mut c_void,
-        fused_out: *mut c_void,
-    ) -> c_int;
-
-    /// Stage A (M3) batched-Q full-attention prefill kernel. Standalone
-    /// attention math: Q/K/V are pre-projected and pre-RoPE'd by the
-    /// caller, the K/V cache is pre-written. Output is pre-o_proj
-    /// `[batch, q_heads, q_len, head_dim]` in F32.
-    ///
-    /// Shapes (BF16 unless noted):
-    /// - `query`: `[batch, q_heads, q_len, head_dim]`
-    /// - `key`:   `[batch, kv_heads, kv_len, head_dim]`
-    /// - `value`: `[batch, kv_heads, kv_len, head_dim]`
-    /// - `out` (F32): `[batch, q_heads, q_len, head_dim]`
-    ///
-    /// `seqlen_offset = past_len`; query at chunk position `qr` attends to
-    /// cache positions `[0, past_len + qr]` (causal, inclusive). `kv_len`
-    /// is the total cache length the kernel may read (typically
-    /// `past_len + q_len`).
-    ///
-    /// Status codes (non-zero = failure):
-    ///   130 dtype != bf16    131 invalid heads        132 q_heads % kv_heads
-    ///   133 head_dim out of range  134 q_len/kv_len   135 seqlen_offset / overflow
-    ///   136 batch_size       137 wave64 (unsupported) 138 LDS overflow
-    ///   254 launch error     255 sync error
-    pub fn qwen36_moe_hip_batched_prefill_attn_full_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        batch_size: c_int,
-        q_heads: c_int,
-        kv_heads: c_int,
-        q_len: c_int,
-        kv_len: c_int,
-        head_dim: c_int,
-        scale: f32,
-        seqlen_offset: c_int,
-        query: *const c_void,
-        key: *const c_void,
-        value: *const c_void,
-        out: *mut c_void,
-    ) -> c_int;
-
-    /// Stage B (M9) router permutation kernel. Groups per-token top-K expert
-    /// assignments by target expert (counting-sort, single block).
-    ///
-    /// Inputs (GPU buffers):
-    /// - `topk_idx`     : `[n_tokens, top_k]` i32 — per-token expert ids in
-    ///                     `[0, num_experts)`.
-    /// - `topk_weight`  : `[n_tokens, top_k]` BF16 — routing weights.
-    ///
-    /// Outputs (caller-allocated GPU buffers):
-    /// - `expert_offsets`     : `[num_experts + 1]` i32 — prefix sum.
-    /// - `permuted_token_idx` : `[n_tokens * top_k]` i32 — sorted token ids.
-    /// - `permuted_kpos`      : `[n_tokens * top_k]` i32 — top-K slot ids.
-    /// - `permuted_weight`    : `[n_tokens * top_k]` BF16 — routing weights.
-    ///
-    /// Within an expert's segment the order is unstable (atomicAdd cursor);
-    /// callers comparing against a CPU reference must compare per-segment as
-    /// a multiset.
-    ///
-    /// Status codes (non-zero = failure):
-    ///   140 invalid args (n_tokens/top_k/num_experts <= 0)
-    ///   141 num_experts > 256       142 top_k > 16
-    ///   143 n_tokens * top_k > 16384
-    ///   254 launch error            255 sync error
-    pub fn qwen36_moe_hip_batched_prefill_router_permute_launch(
-        device_ordinal: usize,
-        n_tokens: c_int,
-        top_k: c_int,
-        num_experts: c_int,
-        topk_idx: *const c_void,
-        topk_weight: *const c_void,
-        expert_offsets: *mut c_void,
-        permuted_token_idx: *mut c_void,
-        permuted_kpos: *mut c_void,
-        permuted_weight: *mut c_void,
-    ) -> c_int;
-
-    /// Stage B (M10) grouped-expert INT4 GEMM kernel. One launch processes
-    /// ALL `num_experts` experts via persistent-block work-stealing on the
-    /// expert id; for each expert it walks the segment of permuted rows
-    /// produced by the M9 router permutation kernel and runs gate_up +
-    /// silu*mul + down INT4 matmuls per row.
-    ///
-    /// Inputs (GPU buffers):
-    /// - `x_norm`              : `[n_tokens, hidden]` BF16 — post-input-RMSnorm
-    ///                            hidden states; gathered by `permuted_token_idx`.
-    /// - `expert_offsets`      : `[num_experts + 1]` i32 — M9 prefix sum.
-    /// - `permuted_token_idx`  : `[n_tokens * top_k]` i32 — M9 sort output.
-    /// - `experts_gate_up_w/s/z` : `[E, 2*I, hidden/2]` u8 + `[E, 2*I/gs, hidden/gs]` BF16.
-    /// - `experts_down_w/s/z`    : `[E, hidden, I/2]` u8 + `[E, hidden/gs, I/gs]` BF16.
-    ///
-    /// Caller-owned buffers:
-    /// - `expert_out` : `[n_tokens * top_k, hidden]` BF16 — per-permuted-row
-    ///                   expert output; M11 unpermutes + combines.
-    /// - `counters`   : `[1]` u32 — work-stealing claim counter; CALLER MUST
-    ///                   ZERO BEFORE LAUNCH.
-    ///
-    /// Status codes (non-zero = failure):
-    ///   150 invalid args (zero/negative dims)
-    ///   151 num_experts > 256
-    ///   152 hidden / moe_intermediate not divisible by group_size (or 16)
-    ///   153 group_size != 128
-    ///   154 top_k * n_tokens > 16384
-    ///   155 dtype != bf16
-    ///   156 LDS overflow
-    ///   254 launch error                255 sync error
-    pub fn qwen36_moe_hip_batched_prefill_grouped_expert_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        n_tokens: c_int,
-        top_k: c_int,
-        num_experts: c_int,
-        hidden: c_int,
-        moe_intermediate: c_int,
-        group_size: c_int,
-        x_norm: *const c_void,
-        expert_offsets: *const c_void,
-        permuted_token_idx: *const c_void,
-        experts_gate_up_w: *const c_void,
-        experts_gate_up_scale: *const c_void,
-        experts_gate_up_zero: *const c_void,
-        experts_down_w: *const c_void,
-        experts_down_scale: *const c_void,
-        experts_down_zero: *const c_void,
-        expert_out: *mut c_void,
-        counters: *mut c_void,
-    ) -> c_int;
-
-    /// Stage B (M11) unpermute + weighted combine kernel. Inverts the M9
-    /// router permutation (host-built `permuted_inverse` table) and computes
-    /// the per-token weighted sum of `top_k` expert outputs.
-    ///
-    /// Inputs (GPU buffers):
-    /// - `permuted_inverse` : `[n_tokens * top_k]` i32 — host-built inverse
-    ///                         of M9's scatter, so
-    ///                         `permuted_inverse[token * top_k + kpos] = dst`
-    ///                         where `dst` is the M9/M10 row index for that
-    ///                         (token, kpos) pair.
-    /// - `permuted_weight`  : `[n_tokens * top_k]` BF16 — M9 output.
-    /// - `expert_out`       : `[n_tokens * top_k, hidden]` BF16 — M10 output.
-    ///
-    /// Output (caller-allocated GPU buffer):
-    /// - `combined`         : `[n_tokens, hidden]` BF16 — weighted sum
-    ///                         of expert outputs per token.
-    ///
-    /// Status codes (non-zero = failure):
-    ///   160 invalid args (zero/negative dims)
-    ///   161 top_k > 16
-    ///   162 dtype != bf16
-    ///   163 hidden too large (>65536)
-    ///   254 launch error            255 sync error
-    pub fn qwen36_moe_hip_batched_prefill_unpermute_combine_launch(
-        dtype: c_int,
-        device_ordinal: usize,
-        n_tokens: c_int,
-        top_k: c_int,
-        hidden: c_int,
-        permuted_inverse: *const c_void,
-        permuted_weight: *const c_void,
-        expert_out: *const c_void,
-        combined: *mut c_void,
-    ) -> c_int;
-}
-
-/// Safe wrapper over the stub launch. The engine pre-allocates `sync_buf`
-/// as a 96-byte zeroed scratch — 16 u32 work-stealing counter slots at
-/// +0..+63 (only counters[0] used here; the FFN concurrent-experts dispatch
-/// uses 2*K_top of them), grid barrier counter at +64, flag at +68. The
-/// 32-byte form used by `crate::persistent_decode_4b` is the older single-
-/// counter layout — qwen36_moe shares one widened sync_buf across all four
-/// step launchers (stub/attn/linear/ffn) so any can run with any.
-///
-/// Returns when the kernel signals completion via `hipDeviceSynchronize`.
-/// The smoke-test path reads `workspace` back to verify descriptor
-/// integrity; the real kernel will overwrite that area with activations.
-pub fn stub_launch(
-    ordinal: usize,
-    dtype: ScalarType,
-    layer_descs_device: &GpuBuffer,
-    workspace: &mut GpuBuffer,
-    sync_buf: &mut GpuBuffer,
-    num_layers: usize,
-) -> Result<(), GpuError> {
-    if dtype != ScalarType::BF16 {
-        return Err(GpuError::InvalidArg(format!(
-            "qwen36_moe::stub_launch: only BF16 is wired, got {dtype:?}"
-        )));
-    }
-    let backend = layer_descs_device.backend();
-    let counters = sync_buf.as_mut_ptr() as *mut c_uint;
-    // Layout: 16 u32 work-stealing counter slots at +0..+63 (the FFN
-    // concurrent-experts dispatch uses 2*K_top of these; attn/linear/stub
-    // only touch counters[0]). Barrier counter+flag follow at +64/+68.
-    // Sync_buf must be at least 96 bytes zeroed before launch.
-    let barrier_counter = unsafe { (counters as *mut u8).add(64) as *mut c_uint };
-    let barrier_flag = unsafe { (counters as *mut u8).add(68) as *mut c_uint };
-
-    let status: c_int = match backend {
-        Backend::Hip | Backend::Cuda => {
-            #[cfg(any(supersonic_backend_hip, supersonic_backend_cuda))]
-            unsafe {
-                qwen36_moe_hip_stub_launch(
-                    dtype.kernel_dtype_code(),
-                    ordinal,
-                    num_layers,
-                    layer_descs_device.as_ptr() as *const Qwen36MoeDecodeLayerDesc,
-                    workspace.as_mut_ptr() as *mut f32,
-                    counters,
-                    barrier_counter,
-                    barrier_flag,
-                )
-            }
-            #[cfg(not(any(supersonic_backend_hip, supersonic_backend_cuda)))]
-            {
-                return Err(GpuError::InvalidArg(
-                    "qwen36_moe::stub_launch: GPU backend not compiled".into(),
-                ));
-            }
-        }
-        Backend::Metal => {
-            return Err(GpuError::InvalidArg(
-                "qwen36_moe::stub_launch: Metal backend not yet wired".into(),
-            ));
-        }
-    };
-    if status != 0 {
-        return Err(GpuError::backend(
-            backend,
-            format!("qwen36_moe stub launch failed with status {status}"),
-        ));
-    }
-    Ok(())
-}
-
-/// Geometry constants for the persistent decode megakernel — packed into
-/// one struct to keep [`persistent_decode_launch`]'s arg list tractable.
-#[derive(Debug, Clone, Copy)]
-pub struct Qwen36MoePersistentGeom {
-    pub hidden: i32,
-    pub num_heads: i32,
-    pub num_kv_heads: i32,
-    pub head_dim: i32,
-    pub rotary_dim: i32,
-    pub num_k_heads: i32,
-    pub num_v_heads: i32,
-    pub head_k_dim: i32,
-    pub head_v_dim: i32,
-    pub conv_kernel_dim: i32,
-    pub num_experts: i32,
-    pub moe_intermediate: i32,
-    pub shared_intermediate: i32,
-    pub top_k: i32,
-    pub rope_theta: f32,
-    pub rms_norm_eps: f32,
-}
-
-/// Phase 3e safe wrapper for the persistent decode megakernel. Replaces
-/// the chained 80 step-kernel launches/token with one cooperative HIP
-/// launch.
-///
-/// Caller responsibilities:
-/// - `layers_device` is a device-resident array of
-///   [`Qwen36MoeDecodeLayerDesc`] (`num_layers` entries). Even
-///   `num_layers` only — the bridge enforces this.
-/// - `int4_scales_device` is null for BF16 bakes, or a device-resident
-///   array of [`Qwen36MoeInt4ScaleDesc`] (parallel to `layers_device`).
-/// - `hidden_ping` is uploaded with the BF16 initial hidden bytes; the
-///   final hidden lands back in `hidden_ping` after `num_layers`.
-/// - `workspace` (F32) sized for `max(attn_workspace_floats(geom),
-///   ffn_workspace_floats(geom))`.
-/// - `ffn_topk_idx_scratch` is a small `[top_k]` i32 buffer (used only
-///   internally by the FFN phase at stage 1; we run stage 5 so it's
-///   inert, but must be valid).
-/// - `sync_buf` is at least 96 zeroed bytes (counters[0..16] + barrier
-///   counter/flag); the bridge defensively re-zeros it on entry.
-/// Phase 3f folded final RMSnorm + lm_head GEMV. Pass `Some(...)` on
-/// generation steps to write logits directly from the megakernel and
-/// skip the separate `lm_head_launch` call; pass `None` on prefill
-/// steps where the caller doesn't need logits. Bundled rather than
-/// scattered as ~4 args so the call site stays readable.
-pub struct Qwen36MoePersistentLmHeadFold<'a> {
-    /// `[hidden]` BF16. Same final_norm tensor the standalone
-    /// `lm_head_launch` consumes.
-    pub final_norm_w: &'a GpuBuffer,
-    /// `[vocab, hidden]` BF16. The bake may store INT4 on disk, but the
-    /// folded kernel reads the pre-dequantized BF16 upload.
-    pub lm_head_w: &'a GpuBuffer,
-    /// Optional `[vocab]` BF16 output buffer. Kernel writes one logit per row
-    /// when full logits are needed for sampling or diagnostics.
-    pub logits_out: Option<&'a mut GpuBuffer>,
-    /// Optional `[1]` U32 output buffer for greedy decode. When present, the
-    /// persistent kernel reduces the lm_head argmax internally.
-    pub top1_out: Option<&'a mut GpuBuffer>,
-    /// Vocab size. Must be `> 0` and match `lm_head_w.shape()[0]`.
-    pub vocab: i32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Qwen36MoePersistentMode {
-    Full,
-    RouterOnly,
-    FfnOnly,
-    AttnOnly,
-    FfnStage(i32),
-    LinearStage(i32),
-}
-
-impl Qwen36MoePersistentMode {
-    #[allow(dead_code)]
-    fn as_ffi(self) -> c_int {
-        match self {
-            Self::Full => 0,
-            Self::RouterOnly => 1,
-            Self::FfnOnly => 2,
-            Self::AttnOnly => 3,
-            Self::FfnStage(stage) => 3 + stage,
-            Self::LinearStage(stage) => 8 + stage,
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn persistent_decode_launch(
-    ordinal: usize,
-    dtype: ScalarType,
-    geom: Qwen36MoePersistentGeom,
-    position: i32,
-    cache_pos: i32,
-    layers_device: &GpuBuffer,
-    int4_scales_device: Option<&GpuBuffer>,
-    kv_fp8_descs_device: Option<&GpuBuffer>,
-    num_layers: usize,
-    hidden_ping: &mut GpuBuffer,
-    hidden_pong: &mut GpuBuffer,
-    workspace: &mut GpuBuffer,
-    ffn_topk_idx_scratch: &mut GpuBuffer,
-    sync_buf: &mut GpuBuffer,
-    lm_head_fold: Option<Qwen36MoePersistentLmHeadFold<'_>>,
-) -> Result<(), GpuError> {
-    persistent_decode_launch_range(
-        ordinal,
-        dtype,
-        geom,
-        0,
-        num_layers,
-        Qwen36MoePersistentMode::Full,
-        position,
-        cache_pos,
-        layers_device,
-        int4_scales_device,
-        kv_fp8_descs_device,
-        num_layers,
-        hidden_ping,
-        hidden_pong,
-        workspace,
-        ffn_topk_idx_scratch,
-        sync_buf,
-        None,
-        -1,
-        None,
-        1,
-        lm_head_fold,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn persistent_decode_launch_range(
-    ordinal: usize,
-    dtype: ScalarType,
-    geom: Qwen36MoePersistentGeom,
-    start_layer: usize,
-    end_layer_exclusive: usize,
-    mode: Qwen36MoePersistentMode,
-    position: i32,
-    // `-1` (`Qwen36MoeAttnStepParams::CACHE_POS_INHERIT`) ⇒ inherit
-    // from `position` (dense base decode). `≥ 0` ⇒ decoupled KV slot
-    // for SpecPrefill sparse-prefill or MTP draft layers.
-    cache_pos: i32,
-    layers_device: &GpuBuffer,
-    int4_scales_device: Option<&GpuBuffer>,
-    kv_fp8_descs_device: Option<&GpuBuffer>,
-    num_layers: usize,
-    hidden_ping: &mut GpuBuffer,
-    hidden_pong: &mut GpuBuffer,
-    workspace: &mut GpuBuffer,
-    ffn_topk_idx_scratch: &mut GpuBuffer,
-    sync_buf: &mut GpuBuffer,
-    embed_w: Option<&GpuBuffer>,
-    token_id: i32,
-    token_ids: Option<&GpuBuffer>,
-    prefill_len: i32,
-    lm_head_fold: Option<Qwen36MoePersistentLmHeadFold<'_>>,
-) -> Result<(), GpuError> {
-    if dtype != ScalarType::BF16 {
-        return Err(GpuError::InvalidArg(format!(
-            "qwen36_moe::persistent_decode_launch: only BF16 wired, got {dtype:?}"
-        )));
-    }
-    if let Qwen36MoePersistentMode::FfnStage(stage) = mode {
-        if !(1..=5).contains(&stage) {
-            return Err(GpuError::InvalidArg(format!(
-                "qwen36_moe::persistent_decode_launch: FFN stage mode must be in 1..=5, got {stage}"
-            )));
-        }
-    }
-    if let Qwen36MoePersistentMode::LinearStage(stage) = mode {
-        if !(1..=5).contains(&stage) {
-            return Err(GpuError::InvalidArg(format!(
-                "qwen36_moe::persistent_decode_launch: linear stage mode must be in 1..=5, got {stage}"
-            )));
-        }
-    }
-    let backend = layers_device.backend();
-    let counters = sync_buf.as_mut_ptr() as *mut c_uint;
-    let barrier_counter = unsafe { (counters as *mut u8).add(64) as *mut c_uint };
-    let barrier_flag = unsafe { (counters as *mut u8).add(68) as *mut c_uint };
-    let int4_ptr: *const Qwen36MoeInt4ScaleDesc = int4_scales_device
-        .map(|b| b.as_ptr() as *const Qwen36MoeInt4ScaleDesc)
-        .unwrap_or(std::ptr::null());
-    let kv_fp8_ptr: *const Qwen36MoeKVCacheFp8Desc = kv_fp8_descs_device
-        .map(|b| b.as_ptr() as *const Qwen36MoeKVCacheFp8Desc)
-        .unwrap_or(std::ptr::null());
-    let embed_w_ptr = embed_w.map(|b| b.as_ptr()).unwrap_or(std::ptr::null());
-    let token_ids_ptr = token_ids
-        .map(|b| b.as_ptr() as *const c_uint)
-        .unwrap_or(std::ptr::null());
-
-    // Fold pointers default to null; the kernel skips the lm_head phase
-    // when any of the three is null.
-    let (vocab, final_norm_w_ptr, lm_head_w_ptr, logits_out_ptr, top1_out_ptr) = match lm_head_fold
-    {
-        Some(mut f) => {
-            let logits_out_ptr = f
-                .logits_out
-                .as_deref_mut()
-                .map(|buf| buf.as_mut_ptr())
-                .unwrap_or(std::ptr::null_mut());
-            let top1_out_ptr = f
-                .top1_out
-                .as_deref_mut()
-                .map(|buf| buf.as_mut_ptr() as *mut c_uint)
-                .unwrap_or(std::ptr::null_mut());
-            (
-                f.vocab,
-                f.final_norm_w.as_ptr(),
-                f.lm_head_w.as_ptr(),
-                logits_out_ptr,
-                top1_out_ptr,
-            )
-        }
-        None => (
-            0,
-            std::ptr::null(),
-            std::ptr::null(),
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-        ),
-    };
-
-    let op = match mode {
-        Qwen36MoePersistentMode::Full => "qwen36.persistent_decode",
-        Qwen36MoePersistentMode::RouterOnly => "qwen36.persistent_router_only",
-        Qwen36MoePersistentMode::FfnOnly => "qwen36.persistent_ffn_only",
-        Qwen36MoePersistentMode::AttnOnly => "qwen36.persistent_attn_only",
-        Qwen36MoePersistentMode::FfnStage(1) => "qwen36.persistent_ffn_stage1",
-        Qwen36MoePersistentMode::FfnStage(2) => "qwen36.persistent_ffn_stage2",
-        Qwen36MoePersistentMode::FfnStage(3) => "qwen36.persistent_ffn_stage3",
-        Qwen36MoePersistentMode::FfnStage(4) => "qwen36.persistent_ffn_stage4",
-        Qwen36MoePersistentMode::FfnStage(5) => "qwen36.persistent_ffn_stage5",
-        Qwen36MoePersistentMode::FfnStage(_) => "qwen36.persistent_ffn_stage_invalid",
-        Qwen36MoePersistentMode::LinearStage(1) => "qwen36.persistent_linear_stage1",
-        Qwen36MoePersistentMode::LinearStage(2) => "qwen36.persistent_linear_stage2",
-        Qwen36MoePersistentMode::LinearStage(3) => "qwen36.persistent_linear_stage3",
-        Qwen36MoePersistentMode::LinearStage(4) => "qwen36.persistent_linear_stage4",
-        Qwen36MoePersistentMode::LinearStage(5) => "qwen36.persistent_linear_stage5",
-        Qwen36MoePersistentMode::LinearStage(_) => "qwen36.persistent_linear_stage_invalid",
-    };
-    crate::prefill_ffi::ffi_profile_time_result(op, ordinal, || {
-        let status: c_int = match backend {
-            Backend::Hip | Backend::Cuda => {
-                #[cfg(any(supersonic_backend_hip, supersonic_backend_cuda))]
-                unsafe {
-                    qwen36_moe_hip_persistent_decode_launch(
-                        dtype.kernel_dtype_code(),
-                        ordinal,
-                        num_layers as c_int,
-                        start_layer as c_int,
-                        end_layer_exclusive as c_int,
-                        mode.as_ffi(),
-                        layers_device.as_ptr() as *const Qwen36MoeDecodeLayerDesc,
-                        int4_ptr,
-                        kv_fp8_ptr,
-                        geom.hidden,
-                        geom.num_heads,
-                        geom.num_kv_heads,
-                        geom.head_dim,
-                        geom.rotary_dim,
-                        geom.num_k_heads,
-                        geom.num_v_heads,
-                        geom.head_k_dim,
-                        geom.head_v_dim,
-                        geom.conv_kernel_dim,
-                        geom.num_experts,
-                        geom.moe_intermediate,
-                        geom.shared_intermediate,
-                        geom.top_k,
-                        vocab,
-                        geom.rope_theta,
-                        geom.rms_norm_eps,
-                        position,
-                        cache_pos,
-                        embed_w_ptr,
-                        token_id as c_int,
-                        token_ids_ptr,
-                        prefill_len as c_int,
-                        hidden_ping.as_mut_ptr(),
-                        hidden_pong.as_mut_ptr(),
-                        workspace.as_mut_ptr() as *mut f32,
-                        ffn_topk_idx_scratch.as_mut_ptr() as *mut c_int,
-                        final_norm_w_ptr,
-                        lm_head_w_ptr,
-                        logits_out_ptr,
-                        top1_out_ptr,
-                        counters,
-                        barrier_counter,
-                        barrier_flag,
-                    )
-                }
-                #[cfg(not(any(supersonic_backend_hip, supersonic_backend_cuda)))]
-                {
-                    return Err(GpuError::InvalidArg(
-                        "qwen36_moe::persistent_decode_launch: GPU backend not compiled".into(),
-                    ));
-                }
-            }
-            Backend::Metal => {
-                return Err(GpuError::InvalidArg(
-                    "qwen36_moe::persistent_decode_launch: Metal backend not yet wired".into(),
-                ));
-            }
-        };
-        if status != 0 {
-            return Err(GpuError::backend(
-                backend,
-                format!("qwen36_moe persistent decode launch failed with status {status}"),
-            ));
-        }
-        Ok(())
-    })
-}
-
 /// Geometry + position state for the staged-attention parity launcher.
 /// These are constants of the layer being tested; bundling them into a
 /// struct keeps the safe wrapper below from sprouting eight scalar args.
